@@ -1,20 +1,18 @@
 {
   "private": true,
   "devDependencies": {
-    "@types/jest": "22.2.2",
-    "@types/node": "9.6.0",
-    "jest": "22.4.3",
-    "lerna": "2.9.0",
-    "ts-jest": "22.4.2",
-    "ts-node": "5.0.1",
-    "tslint": "5.9.1",
-    "typescript": "2.7.2"
+    "@types/jest": "^26.0.21",
+    "@types/node": "^14.14.35",
+    "jest": "^26.6.3",
+    "lerna": "^4.0.0",
+    "ts-jest": "^26.5.4",
+    "ts-node": "^9.1.1",
+    "tslint": "^6.1.3",
+    "typescript": "^4.2.3"
   },
   "scripts": {
     "reset": "rm -rf node_modules yarn.lock packages/**/yarn.lock && yarn && yarn all",
-    "docs": "packages/oly-docs/bin/oly-docs",
-    "docs:dev": "HTML=false OUT=packages/oly-docs/src/client/assets yarn docs --verbose",
-    "all": "lerna run clean && lerna bootstrap && lerna run build && yarn test:ci && yarn docs",
+    "all": "lerna run clean && lerna bootstrap && lerna run build && yarn test:ci",
     "clean": "lerna run clean",
     "bootstrap": "lerna bootstrap",
     "watch": "lerna exec --parallel -- npm run watch",
